[99]Call DLL in Python 2

+++ 99特色功能/Call DLL in Python/ConsoleApplication2/ConsoleApplication2/notification.hpp
@@ -2,8 +2,7 @@
 #define NOTIFICATION_HPP
 
 #include <vector>
-#include "warningresourceids.hpp"
-#include "warningstrategy.hpp"
+#include "typedefine.hpp"
 
 // The template component.
 class Notification
@@ -39,7 +38,7 @@ public:
 		int m_ACID;
 	};
 
-    unsigned int u16ActiveNotificationID;
+	uint16 u16ActiveNotificationID;
